move and update the `about` page
